+++ myec3-socle-contract/src/main/java/org/myec3/socle/core/domain/model/Structure.java
@@ -18,6 +18,7 @@
 package org.myec3.socle.core.domain.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.LazyCollection;
@@ -75,6 +76,9 @@ public abstract class Structure extends Resource {
 	private StructureType structureType;
 	private Integer workforce;
 	private String tenantIdentifier;
+
+	@Transient
+	private CompanyLogo companyLogo;
 
 	@Getter(onMethod=@__(@XmlTransient))
 	@Column(nullable = false)
@@ -386,6 +390,18 @@ public abstract class Structure extends Resource {
 	public void setLogoUrl(String logoUrl) {
 		this.logoUrl = logoUrl;
 	}
+
+	@XmlElement(required = false)
+	@JsonProperty("company_logo")
+	@Transient
+	public CompanyLogo getCompanyLogo() {
+		return companyLogo;
+	}
+
+	public void setPays(CompanyLogo companyLogo) {
+		this.companyLogo = companyLogo;
+	}
+
 
 	/**
 	 * Address of this structure. Should be a valid address object. Cannot be null.