+++ myec3-socle-synchro-module/myec3-socle-synchro/src/main/java/org/myec3/socle/synchro/controller/SynchroController.java
@@ -1,5 +1,6 @@
 package org.myec3.socle.synchro.controller;
 
+import org.myec3.socle.core.constants.MyEc3ApplicationConstants;
 import org.myec3.socle.core.domain.dto.OrganismLightDTO;
 import org.myec3.socle.core.domain.model.*;
 import org.myec3.socle.core.service.*;
@@ -93,14 +94,13 @@ public class SynchroController {
 
         AgentProfile agent = agentProfileService.findOne(id);
 
-        Application applicationASynchroniser = applicationService.findByName("portail megalisbretagne");
+        Application applicationASynchroniser = applicationService.findByName(MyEc3ApplicationConstants.PORTAIL_MEGALIS_APPLICATION);
         List<Long> listApplicationIdToResynchronize = new ArrayList<>();
         listApplicationIdToResynchronize.add(applicationASynchroniser.getId());
 
         SynchronizationType synchronizationType = SynchronizationType.SYNCHRONIZATION;
-        String sendingApplication = "GU";
 
-        agentSynchronizer.synchronizeUpdate(agent, listApplicationIdToResynchronize, synchronizationType, sendingApplication);
+        agentSynchronizer.synchronizeUpdate(agent, listApplicationIdToResynchronize, synchronizationType, MyEc3ApplicationConstants.GU);
         logger.info("synchro agent {}", agent.getUsername());
 
         return "synchro agent " + agent.getName();
@@ -112,14 +112,13 @@ public class SynchroController {
 
         AgentProfile agent = agentProfileService.findOne(id);
 
-        Application applicationASynchroniser = applicationService.findByName("Pastell");
+        Application applicationASynchroniser = applicationService.findByName(MyEc3ApplicationConstants.PASTELL_APPLICATION);
         List<Long> listApplicationIdToResynchronize = new ArrayList<>();
         listApplicationIdToResynchronize.add(applicationASynchroniser.getId());
 
         SynchronizationType synchronizationType = SynchronizationType.SYNCHRONIZATION;
-        String sendingApplication = "GU";
 
-        agentSynchronizer.synchronizeUpdate(agent, listApplicationIdToResynchronize, synchronizationType, sendingApplication);
+        agentSynchronizer.synchronizeUpdate(agent, listApplicationIdToResynchronize, synchronizationType, MyEc3ApplicationConstants.GU);
         logger.info("synchro agent {}", agent.getUsername());
 
         return "synchro agent " + agent.getName();
@@ -129,13 +128,12 @@ public class SynchroController {
     public String organismePastell(@RequestParam long id) {
         Organism organism = organismService.findOne(id);
 
-        Application applicationASynchroniser = applicationService.findByName("Pastell");
+        Application applicationASynchroniser = applicationService.findByName(MyEc3ApplicationConstants.PASTELL_APPLICATION);
         List<Long> listApplicationIdToResynchronize = new ArrayList<>();
         listApplicationIdToResynchronize.add(applicationASynchroniser.getId());
         SynchronizationType synchronizationType = SynchronizationType.SYNCHRONIZATION;
-        String sendingApplication = "GU";
 
-        organismSynchronizer.synchronizeUpdate(organism, listApplicationIdToResynchronize, synchronizationType, sendingApplication);
+        organismSynchronizer.synchronizeUpdate(organism, listApplicationIdToResynchronize, synchronizationType, MyEc3ApplicationConstants.GU);
 
         return "synchro organism " + organism.getName();
     }
@@ -145,13 +143,12 @@ public class SynchroController {
     public String organismDepartmentPastell(@RequestParam long id) {
         OrganismDepartment organismDepartment = organismDepartmentService.findOne(id);
 
-        Application applicationASynchroniser = applicationService.findByName("Pastell");
+        Application applicationASynchroniser = applicationService.findByName(MyEc3ApplicationConstants.PASTELL_APPLICATION);
         List<Long> listApplicationIdToResynchronize = new ArrayList<>();
         listApplicationIdToResynchronize.add(applicationASynchroniser.getId());
         SynchronizationType synchronizationType = SynchronizationType.SYNCHRONIZATION;
-        String sendingApplication = "GU";
 
-        organismDepartmentSynchronizer.synchronizeUpdate(organismDepartment, listApplicationIdToResynchronize, synchronizationType, sendingApplication);
+        organismDepartmentSynchronizer.synchronizeUpdate(organismDepartment, listApplicationIdToResynchronize, synchronizationType, MyEc3ApplicationConstants.GU);
 
         return "synchro organismDepartment " + organismDepartment.getName();
     }
@@ -162,12 +159,12 @@ public class SynchroController {
 
         AgentProfile agent = agentProfileService.findOne(id);
 
-        Application applicationASynchroniser = applicationService.findByName("SDM");
+        Application applicationASynchroniser = applicationService.findByName(MyEc3ApplicationConstants.SDM_APPLICATION);
         List<Long> listApplicationIdToResynchronize = new ArrayList<>();
         listApplicationIdToResynchronize.add(applicationASynchroniser.getId());
 
         SynchronizationType synchronizationType = SynchronizationType.SYNCHRONIZATION;
-        String sendingApplication = "GU";
+        String sendingApplication = MyEc3ApplicationConstants.GU;
 
         agentSynchronizer.synchronizeUpdate(agent, listApplicationIdToResynchronize, synchronizationType, sendingApplication);
         logger.info("synchro agent {}", agent.getUsername());
@@ -179,13 +176,12 @@ public class SynchroController {
     public String organisme(@RequestParam long id) {
         Organism organism = organismService.findOne(id);
 
-        Application applicationASynchroniser = applicationService.findByName("SDM");
+        Application applicationASynchroniser = applicationService.findByName(MyEc3ApplicationConstants.SDM_APPLICATION);
         List<Long> listApplicationIdToResynchronize = new ArrayList<>();
         listApplicationIdToResynchronize.add(applicationASynchroniser.getId());
         SynchronizationType synchronizationType = SynchronizationType.SYNCHRONIZATION;
-        String sendingApplication = "GU";
 
-        organismSynchronizer.synchronizeUpdate(organism, listApplicationIdToResynchronize, synchronizationType, sendingApplication);
+        organismSynchronizer.synchronizeUpdate(organism, listApplicationIdToResynchronize, synchronizationType, MyEc3ApplicationConstants.GU);
 
         return "synchro organism " + organism.getName();
     }
@@ -194,13 +190,12 @@ public class SynchroController {
     public String organismDepartment(@RequestParam long id) {
         OrganismDepartment organismDepartment = organismDepartmentService.findOne(id);
 
-        Application applicationASynchroniser = applicationService.findByName("SDM");
+        Application applicationASynchroniser = applicationService.findByName(MyEc3ApplicationConstants.SDM_APPLICATION);
         List<Long> listApplicationIdToResynchronize = new ArrayList<>();
         listApplicationIdToResynchronize.add(applicationASynchroniser.getId());
         SynchronizationType synchronizationType = SynchronizationType.SYNCHRONIZATION;
-        String sendingApplication = "GU";
 
-        organismDepartmentSynchronizer.synchronizeUpdate(organismDepartment, listApplicationIdToResynchronize, synchronizationType, sendingApplication);
+        organismDepartmentSynchronizer.synchronizeUpdate(organismDepartment, listApplicationIdToResynchronize, synchronizationType, MyEc3ApplicationConstants.GU);
 
         return "synchro organismDepartment " + organismDepartment.getName();
     }
@@ -210,13 +205,12 @@ public class SynchroController {
     public String employee(@RequestParam long id) {
         EmployeeProfile employeeProfile = employeeProfileService.findOne(id);
 
-        Application applicationASynchroniser = applicationService.findByName("SDM");
+        Application applicationASynchroniser = applicationService.findByName(MyEc3ApplicationConstants.SDM_APPLICATION);
         List<Long> listApplicationIdToResynchronize = new ArrayList<>();
         listApplicationIdToResynchronize.add(applicationASynchroniser.getId());
         SynchronizationType synchronizationType = SynchronizationType.SYNCHRONIZATION;
-        String sendingApplication = "GU";
 
-        employeeSynchronizer.synchronizeUpdate(employeeProfile, listApplicationIdToResynchronize, synchronizationType, sendingApplication);
+        employeeSynchronizer.synchronizeUpdate(employeeProfile, listApplicationIdToResynchronize, synchronizationType, MyEc3ApplicationConstants.GU);
 
         return "synchro employee " + employeeProfile.getName();
     }
@@ -225,13 +219,12 @@ public class SynchroController {
     public String company(@RequestParam long id) {
         Company company = companyService.findOne(id);
 
-        Application applicationASynchroniser = applicationService.findByName("SDM");
+        Application applicationASynchroniser = applicationService.findByName(MyEc3ApplicationConstants.SDM_APPLICATION);
         List<Long> listApplicationIdToResynchronize = new ArrayList<>();
         listApplicationIdToResynchronize.add(applicationASynchroniser.getId());
         SynchronizationType synchronizationType = SynchronizationType.SYNCHRONIZATION;
-        String sendingApplication = "GU";
 
-        companySynchronizer.synchronizeUpdate(company, listApplicationIdToResynchronize, synchronizationType, sendingApplication);
+        companySynchronizer.synchronizeUpdate(company, listApplicationIdToResynchronize, synchronizationType, MyEc3ApplicationConstants.GU);
 
         return "synchro company " + company.getName();
     }
@@ -240,13 +233,12 @@ public class SynchroController {
     public String establishment(@RequestParam long id) {
         Establishment establishment = establishmentService.findOne(id);
 
-        Application applicationASynchroniser = applicationService.findByName("SDM");
+        Application applicationASynchroniser = applicationService.findByName(MyEc3ApplicationConstants.SDM_APPLICATION);
         List<Long> listApplicationIdToResynchronize = new ArrayList<>();
         listApplicationIdToResynchronize.add(applicationASynchroniser.getId());
         SynchronizationType synchronizationType = SynchronizationType.SYNCHRONIZATION;
-        String sendingApplication = "GU";
 
-        establishmentSynchronizer.synchronizeUpdate(establishment, listApplicationIdToResynchronize, synchronizationType, sendingApplication);
+        establishmentSynchronizer.synchronizeUpdate(establishment, listApplicationIdToResynchronize, synchronizationType, MyEc3ApplicationConstants.GU);
 
         return "synchro establishment " + establishment.getName();
     }
@@ -255,13 +247,12 @@ public class SynchroController {
     public String udataOrganism(@RequestParam long id) {
         Organism organism = organismService.findOne(id);
 
-        Application applicationASynchroniser = applicationService.findByName("Udata");
+        Application applicationASynchroniser = applicationService.findByName(MyEc3ApplicationConstants.UDATA_APPLICATION);
         List<Long> listApplicationIdToResynchronize = new ArrayList<>();
         listApplicationIdToResynchronize.add(applicationASynchroniser.getId());
         SynchronizationType synchronizationType = SynchronizationType.SYNCHRONIZATION;
-        String sendingApplication = "GU";
 
-        organismSynchronizer.synchronizeUpdate(organism, listApplicationIdToResynchronize, synchronizationType, sendingApplication);
+        organismSynchronizer.synchronizeUpdate(organism, listApplicationIdToResynchronize, synchronizationType, MyEc3ApplicationConstants.GU);
 
         return "synchro organism " + organism.getName();
     }
@@ -270,13 +261,12 @@ public class SynchroController {
     public String udataAgent(@RequestParam long id) {
         AgentProfile agentProfile = agentProfileService.findOne(id);
 
-        Application applicationASynchroniser = applicationService.findByName("Udata");
+        Application applicationASynchroniser = applicationService.findByName(MyEc3ApplicationConstants.UDATA_APPLICATION);
         List<Long> listApplicationIdToResynchronize = new ArrayList<>();
         listApplicationIdToResynchronize.add(applicationASynchroniser.getId());
         SynchronizationType synchronizationType = SynchronizationType.SYNCHRONIZATION;
-        String sendingApplication = "GU";
 
-        agentSynchronizer.synchronizeUpdate(agentProfile, listApplicationIdToResynchronize, synchronizationType, sendingApplication);
+        agentSynchronizer.synchronizeUpdate(agentProfile, listApplicationIdToResynchronize, synchronizationType, MyEc3ApplicationConstants.GU);
 
         return "synchro agent " + agentProfile.getName();
     }
@@ -293,7 +283,6 @@ public class SynchroController {
         logger.info("[RESYNC] Call resync Organism");
 
         SynchronizationType synchronizationType = SynchronizationType.SYNCHRONIZATION;
-        String sendingApplication = "GU";
 
         // Get all organism subscribe to SDM
         List<OrganismLightDTO> organisms = organismService.findOrganismLight()
@@ -324,7 +313,7 @@ public class SynchroController {
                         Organism organism = organismService.findOne(organismLightDTO.getId());
                         organism.setLabel(labelInsee);
                         organism = organismService.update(organism);
-                        organismSynchronizer.synchronizeUpdate(organism, null, synchronizationType, sendingApplication);
+                        organismSynchronizer.synchronizeUpdate(organism, null, synchronizationType, MyEc3ApplicationConstants.GU);
                         logger.info("[RESYNC] Organism " + organismLightDTO.getId() + " Updated");
                     }
                 }