+++ myec3-socle-webapp-module/myec3-socle-webapp-core/src/main/java/org/myec3/socle/webapp/pages/organism/department/ListAgents.java
@@ -19,14 +19,14 @@ package org.myec3.socle.webapp.pages.organism.department;
 
 import org.apache.tapestry5.EventConstants;
 import org.apache.tapestry5.PersistenceConstants;
-import org.apache.tapestry5.PropertyConduit;
+import org.apache.tapestry5.beanmodel.PropertyConduit;
 import org.apache.tapestry5.annotations.OnEvent;
 import org.apache.tapestry5.annotations.Persist;
 import org.apache.tapestry5.annotations.Property;
-import org.apache.tapestry5.beaneditor.BeanModel;
+import org.apache.tapestry5.beanmodel.BeanModel;
 import org.apache.tapestry5.ioc.annotations.Inject;
-import org.apache.tapestry5.services.BeanModelSource;
-import org.apache.tapestry5.services.PropertyConduitSource;
+import org.apache.tapestry5.beanmodel.services.BeanModelSource;
+import org.apache.tapestry5.beanmodel.services.PropertyConduitSource;
 import org.apache.tapestry5.services.javascript.JavaScriptSupport;
 import org.myec3.socle.core.domain.model.AgentProfile;
 import org.myec3.socle.core.domain.model.Organism;
@@ -195,17 +195,17 @@ public class ListAgents extends AbstractPage {
 	 */
 	public BeanModel<AgentProfile> getGridModel() {
 		BeanModel<AgentProfile> model = this.beanModelSource.createDisplayModel(AgentProfile.class, this.getMessages());
-		model.add("actions", null);
 
-		PropertyConduit propCdtAttributeLastName = this.propertyConduitSource.create(AgentProfile.class,
-				"user.lastname");
-		PropertyConduit propCdtAttributeExpirationDatePassword = this.propertyConduitSource.create(AgentProfile.class,
-				"user.expirationDatePassword");
+		PropertyConduit propCdtAttributeLastName = this.propertyConduitSource.create(AgentProfile.class, "user.lastname");
+		PropertyConduit propCdtAttributeExpirationDatePassword = this.propertyConduitSource.create(AgentProfile.class, "user.expirationDatePassword");
 
 		model.add("user", propCdtAttributeLastName).sortable(true);
 		model.add("expirationDatePassword", propCdtAttributeExpirationDatePassword).sortable(true);
-		model.include("user", "email", "expirationDatePassword", "actions");
+		model.add("actions", null);
+		model.include("user", "email", "username", "expirationDatePassword", "actions");
 		return model;
+
+
 	}
 
 	public OrganismDepartment getOrganismDepartment() {