+++ CHANGELOG.md
@@ -7,7 +7,8 @@ Basé sur [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 - Ajout script des scripts d'initialisation de la base de données
-- 
+- [MEGALIS-292] (https://dev.sib.fr/bts/browse/MEGALIS-292): Ajouter le lien de régénération de mot de passe sur mail d'expiration
+- [MEGALIS-295] (https://dev.sib.fr/bts/browse/MEGALIS-295): Ajouter de la date de création du compte dans l'export des agents effectué depuis le socle
 
 ## [0.2.7] - 2023-02-07